Revert "Merge branch 'delivery/release_user' into pj/at/dev/team_G/KDM001"
This reverts commit f63b4d004b422fbfc92f19a576f34cdb09cfceb8, reversing
changes made to cb7427fe2a5bda554f016226f5da7282934a347b.

+++ nts.uk/uk.at/at.ctx/function/nts.uk.ctx.at.function.pub/src/main/java/nts/uk/ctx/at/function/pub/alarm/checkcondition/eventdto/SpecHolidayCheckConAdapterPubDto.java
@@ -1,7 +1,5 @@
 package nts.uk.ctx.at.function.pub.alarm.checkcondition.eventdto;
 
-import java.math.BigDecimal;
-
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -15,10 +13,10 @@ public class SpecHolidayCheckConAdapterPubDto {
 	/**比較演算子*/
 	private int compareOperator; 
 	/**所定公休日数との差分の日数1*/
-	private BigDecimal numberDayDiffHoliday1;
+	private int numberDayDiffHoliday1;
 	/**所定公休日数との差分の日数2*/
-	private BigDecimal numberDayDiffHoliday2;
-	public SpecHolidayCheckConAdapterPubDto(String errorAlarmCheckID, int compareOperator, BigDecimal numberDayDiffHoliday1, BigDecimal numberDayDiffHoliday2) {
+	private Integer numberDayDiffHoliday2;
+	public SpecHolidayCheckConAdapterPubDto(String errorAlarmCheckID, int compareOperator, int numberDayDiffHoliday1, Integer numberDayDiffHoliday2) {
 		super();
 		this.errorAlarmCheckID = errorAlarmCheckID;
 		this.compareOperator = compareOperator;
